Replacing the Optimize.java with Optimize.kt
