+++ code/optimize/src/main/java/com/adobe/marketing/mobile/optimize/Optimize.kt
@@ -1,5 +1,5 @@
 /*
-  Copyright 2021 Adobe. All rights reserved.
+  Copyright 2025 Adobe. All rights reserved.
   This file is licensed to you under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License. You may obtain a copy
   of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,158 +9,306 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.optimize;
+package com.adobe.marketing.mobile.optimize
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import com.adobe.marketing.mobile.AdobeCallback;
-import com.adobe.marketing.mobile.Extension;
-import java.util.List;
-import java.util.Map;
+import com.adobe.marketing.mobile.AdobeCallback
+import com.adobe.marketing.mobile.AdobeCallbackWithError
+import com.adobe.marketing.mobile.AdobeError
+import com.adobe.marketing.mobile.Event
+import com.adobe.marketing.mobile.Extension
+import com.adobe.marketing.mobile.MobileCore
+import com.adobe.marketing.mobile.services.Log
+import com.adobe.marketing.mobile.util.DataReader
+import com.adobe.marketing.mobile.util.DataReaderException
 
-/** Public class containing APIs for the Optimize extension. */
-public class Optimize {
-    public static final Class<? extends Extension> EXTENSION = OptimizeExtension.class;
+object Optimize {
 
-    private Optimize() {}
+    private const val SELF_TAG: String = "Optimize"
 
-    /**
-     * Returns the version of the {@code Optimize} extension.
-     *
-     * @return {@link String} containing the current installed version of this extension.
-     */
-    @NonNull public static String extensionVersion() {
-        return OptimizeConstants.EXTENSION_VERSION;
+    @JvmField
+    val EXTENSION: Class<out Extension?> = OptimizeExtension::class.java
+
+    @JvmStatic
+    fun extensionVersion(): String = OptimizeConstants.EXTENSION_VERSION
+
+    @JvmStatic
+    fun onPropositionsUpdate(callback: AdobeCallback<Map<DecisionScope, OptimizeProposition>>) {
+        MobileCore.registerEventListener(
+            OptimizeConstants.EventType.OPTIMIZE,
+            OptimizeConstants.EventSource.NOTIFICATION,
+            object : AdobeCallbackWithError<Event> {
+                override fun fail(error: AdobeError) {
+                }
+
+                override fun call(event: Event) {
+                    val eventData = event.eventData ?: return
+
+                    try {
+                        val propositionsList = DataReader.getTypedListOfMap(
+                            Any::class.java,
+                            eventData,
+                            OptimizeConstants.EventDataKeys.PROPOSITIONS
+                        )
+
+                        val propositionsMap = mutableMapOf<DecisionScope, OptimizeProposition>()
+                        propositionsList?.forEach { propositionData ->
+                            val optimizeProposition =
+                                OptimizeProposition.fromEventData(propositionData)
+                            if (optimizeProposition != null && !optimizeProposition.scope.isNullOrEmpty()) {
+                                val scope = DecisionScope(optimizeProposition.scope)
+                                propositionsMap[scope] = optimizeProposition
+                            }
+                        }
+
+                        if (propositionsMap.isNotEmpty()) {
+                            callback.call(propositionsMap)
+                        }
+                    } catch (ignored: DataReaderException) {
+                    }
+                }
+            }
+        )
     }
 
-    /**
-     * This API dispatches an Event for the Edge network extension to fetch decision propositions,
-     * for the provided decision scopes list, from the decisioning services enabled in the
-     * Experience Edge network.
-     *
-     * <p>The returned decision propositions are cached in-memory in the Optimize SDK extension and
-     * can be retrieved using {@link #getPropositions(List, AdobeCallback)} API.
-     *
-     * @param decisionScopes {@code List<DecisionScope>} containing scopes for which offers need to
-     *     be updated.
-     * @param xdm {@code Map<String, Object>} containing additional XDM-formatted data to be sent in
-     *     the personalization query request.
-     * @param data {@code Map<String, Object>} containing additional free-form data to be sent in
-     *     the personalization query request.
-     */
-    @Deprecated
-    public static void updatePropositions(
-            @NonNull final List<DecisionScope> decisionScopes,
-            @Nullable final Map<String, Object> xdm,
-            @Nullable final Map<String, Object> data) {
-
-        updatePropositions(decisionScopes, xdm, data, null);
+    @JvmStatic
+    fun clearCachedPropositions() {
+        val event = Event.Builder(
+            OptimizeConstants.EventNames.CLEAR_PROPOSITIONS_REQUEST,
+            OptimizeConstants.EventType.OPTIMIZE,
+            OptimizeConstants.EventSource.REQUEST_RESET
+        ).build()
+        MobileCore.dispatchEvent(event)
     }
 
-    /**
-     * This API dispatches an Event for the Edge network extension to fetch decision propositions,
-     * for the provided decision scopes list, from the decisioning services enabled in the
-     * Experience Edge network.
-     *
-     * <p>The returned decision propositions are cached in-memory in the Optimize SDK extension and
-     * can be retrieved using {@link #getPropositions(List, AdobeCallback)} API.
-     *
-     * @param decisionScopes {@code List<DecisionScope>} containing scopes for which offers need to
-     *     be updated.
-     * @param xdm {@code Map<String, Object>} containing additional XDM-formatted data to be sent in
-     *     the personalization query request.
-     * @param data {@code Map<String, Object>} containing additional free-form data to be sent in
-     *     the personalization query request.
-     * @param callback {@code AdobeCallback<Map<DecisionScope, OptimizeProposition>>} which will be
-     *     invoked when decision propositions are received from the Edge network.
-     */
-    public static void updatePropositions(
-            @NonNull final List<DecisionScope> decisionScopes,
-            @Nullable final Map<String, Object> xdm,
-            @Nullable final Map<String, Object> data,
-            @Nullable final AdobeCallback<Map<DecisionScope, OptimizeProposition>> callback) {
-        OptimizeImpl.INSTANCE.updatePropositionsInternal(decisionScopes, xdm, data, callback);
+    @JvmStatic
+    fun getPropositions(
+        decisionScopes: List<DecisionScope>,
+        callback: AdobeCallback<Map<DecisionScope, OptimizeProposition>>
+    ){
+        getPropositions(decisionScopes, OptimizeConstants.GET_RESPONSE_CALLBACK_TIMEOUT, callback)
     }
 
-    /**
-     * This API dispatches an Event for the Edge network extension to fetch decision propositions,
-     * for the provided decision scopes list, from the decisioning services enabled in the
-     * Experience Edge network.
-     *
-     * <p>The returned decision propositions are cached in-memory in the Optimize SDK extension and
-     * can be retrieved using {@link #getPropositions(List, double, AdobeCallback)} API.
-     *
-     * @param decisionScopes {@code List<DecisionScope>} containing scopes for which offers need to
-     *     be updated.
-     * @param xdm {@code Map<String, Object>} containing additional XDM-formatted data to be sent in
-     *     the personalization query request.
-     * @param data {@code Map<String, Object>} containing additional free-form data to be sent in
-     *     the personalization query request.
-     * @param timeoutSeconds {@code Double} containing additional configurable timeout(seconds) to
-     *     be sent in the personalization query request.
-     * @param callback {@code AdobeCallback<Map<DecisionScope, OptimizeProposition>>} which will be
-     *     invoked when decision propositions are received from the Edge network.
-     */
-    public static void updatePropositions(
-            @NonNull final List<DecisionScope> decisionScopes,
-            @Nullable final Map<String, Object> xdm,
-            @Nullable final Map<String, Object> data,
-            final double timeoutSeconds,
-            @Nullable final AdobeCallback<Map<DecisionScope, OptimizeProposition>> callback) {
-        OptimizeImpl.INSTANCE.updatePropositionsInternal(
-                decisionScopes, xdm, data, timeoutSeconds, callback);
+    @JvmStatic
+    fun getPropositions(
+        decisionScopes: List<DecisionScope>,
+        timeoutSeconds: Double = OptimizeConstants.GET_RESPONSE_CALLBACK_TIMEOUT,
+        callback: AdobeCallback<Map<DecisionScope, OptimizeProposition>>
+    ) {
+        if (decisionScopes.isNullOrEmpty()) {
+            Log.warning(
+                OptimizeConstants.LOG_TAG,
+                SELF_TAG,
+                "Cannot get propositions, provided list of decision scopes is null or empty."
+            )
+            OptimizeUtils.failWithError(callback, AdobeError.UNEXPECTED_ERROR)
+            return
+        }
+
+        val validScopes = decisionScopes.filter { it.isValid() }
+
+        if (validScopes.isEmpty()) {
+            Log.warning(
+                OptimizeConstants.LOG_TAG,
+                SELF_TAG,
+                "Cannot get propositions, provided list of decision scopes has no valid scope."
+            )
+            OptimizeUtils.failWithError(callback, AdobeError.UNEXPECTED_ERROR)
+            return
+        }
+
+        val flattenedDecisionScopes = validScopes.map { it.toEventData() }
+
+        val eventData = mutableMapOf<String, Any>().apply {
+            put(
+                OptimizeConstants.EventDataKeys.REQUEST_TYPE,
+                OptimizeConstants.EventDataValues.REQUEST_TYPE_GET
+            )
+            put(OptimizeConstants.EventDataKeys.DECISION_SCOPES, flattenedDecisionScopes)
+        }
+
+        val event = Event.Builder(
+            OptimizeConstants.EventNames.GET_PROPOSITIONS_REQUEST,
+            OptimizeConstants.EventType.OPTIMIZE,
+            OptimizeConstants.EventSource.REQUEST_CONTENT
+        ).setEventData(eventData).build()
+
+        val timeoutMillis = (timeoutSeconds * OptimizeConstants.TIMEOUT_CONVERSION_FACTOR).toLong()
+
+        MobileCore.dispatchEventWithResponseCallback(
+            event,
+            timeoutMillis,
+            object : AdobeCallbackWithError<Event> {
+                override fun fail(adobeError: AdobeError) {
+                    OptimizeUtils.failWithError(callback, adobeError)
+                }
+
+                override fun call(event: Event) {
+                    try {
+                        val eventData = event.eventData ?: run {
+                            OptimizeUtils.failWithError(callback, AdobeError.UNEXPECTED_ERROR)
+                            return
+                        }
+
+                        if (eventData.containsKey(OptimizeConstants.EventDataKeys.RESPONSE_ERROR)) {
+                            val errorCode = DataReader.getInt(
+                                eventData,
+                                OptimizeConstants.EventDataKeys.RESPONSE_ERROR
+                            )
+                            OptimizeUtils.failWithError(
+                                callback,
+                                OptimizeUtils.convertToAdobeError(errorCode)
+                            )
+                            return
+                        }
+
+                        val propositionsList = DataReader.getTypedListOfMap(
+                            Any::class.java,
+                            eventData,
+                            OptimizeConstants.EventDataKeys.PROPOSITIONS
+                        )
+
+                        val propositionsMap = mutableMapOf<DecisionScope, OptimizeProposition>()
+                        propositionsList?.forEach { propositionData ->
+                            val optimizeProposition =
+                                OptimizeProposition.fromEventData(propositionData)
+                            if (optimizeProposition != null && !optimizeProposition.scope.isNullOrEmpty()) {
+                                val scope = DecisionScope(optimizeProposition.scope)
+                                propositionsMap[scope] = optimizeProposition
+                            }
+                        }
+
+                        callback.call(propositionsMap)
+                    } catch (e: DataReaderException) {
+                        OptimizeUtils.failWithError(callback, AdobeError.UNEXPECTED_ERROR)
+                    }
+                }
+            }
+        )
     }
 
-    /**
-     * This API retrieves the previously fetched propositions, for the provided decision scopes,
-     * from the in-memory extension propositions cache.
-     *
-     * @param decisionScopes {@code List<DecisionScope>} containing scopes for which offers need to
-     *     be requested.
-     * @param callback {@code AdobeCallbackWithError<Map<DecisionScope, OptimizeProposition>>} which
-     *     will be invoked when decision propositions are retrieved from the local cache.
-     */
-    public static void getPropositions(
-            @NonNull final List<DecisionScope> decisionScopes,
-            @NonNull final AdobeCallback<Map<DecisionScope, OptimizeProposition>> callback) {
-        OptimizeImpl.INSTANCE.getPropositionsInternal(decisionScopes, callback);
+    @JvmStatic
+    @JvmOverloads
+    fun updatePropositions(
+        decisionScopes: List<DecisionScope>,
+        xdm: Map<String, Any>? = null,
+        data: Map<String, Any>? = null,
+        callback: AdobeCallback<Map<DecisionScope, OptimizeProposition>>
+    ){
+        updatePropositions(decisionScopes, xdm, data, OptimizeConstants.EDGE_CONTENT_COMPLETE_RESPONSE_TIMEOUT, callback)
     }
 
-    /**
-     * This API retrieves the previously fetched propositions, for the provided decision scopes,
-     * from the in-memory extension propositions cache.
-     *
-     * @param decisionScopes {@code List<DecisionScope>} containing scopes for which offers need to
-     *     be requested.
-     * @param callback {@code AdobeCallbackWithError<Map<DecisionScope, OptimizeProposition>>} which
-     *     will be invoked when decision propositions are retrieved from the local cache.
-     */
-    public static void getPropositions(
-            @NonNull final List<DecisionScope> decisionScopes,
-            final double timeoutSeconds,
-            @NonNull final AdobeCallback<Map<DecisionScope, OptimizeProposition>> callback) {
-        OptimizeImpl.INSTANCE.getPropositionsInternal(decisionScopes, timeoutSeconds, callback);
-    }
+    @JvmStatic
+    fun updatePropositions(
+        decisionScopes: List<DecisionScope>,
+        xdm: Map<String, Any>? = null,
+        data: Map<String, Any>? = null,
+        timeoutSeconds: Double = OptimizeConstants.EDGE_CONTENT_COMPLETE_RESPONSE_TIMEOUT,
+        callback: AdobeCallback<Map<DecisionScope, OptimizeProposition>>
+    ) {
+        if (decisionScopes.isNullOrEmpty()) {
+            Log.warning(
+                OptimizeConstants.LOG_TAG,
+                SELF_TAG,
+                "Cannot update propositions, provided list of decision scopes is null or empty."
+            )
 
-    /**
-     * This API registers a permanent callback which is invoked whenever the Edge extension
-     * dispatches a response Event received from the Experience Edge Network upon a personalization
-     * query.
-     *
-     * <p>The personalization query requests can be triggered by the {@link
-     * Optimize#updatePropositions(List, Map, Map)} API, Edge extension {@code
-     * sendEvent(ExperienceEvent, EdgeCallback)} API or launch consequence rules.
-     *
-     * @param callback {@code AdobeCallbackWithError<Map<DecisionScope, OptimizeProposition>>} which
-     *     will be invoked when decision propositions are received from the Edge network.
-     */
-    public static void onPropositionsUpdate(
-            @NonNull final AdobeCallback<Map<DecisionScope, OptimizeProposition>> callback) {
-        OptimizeImpl.INSTANCE.onPropositionsUpdate(callback);
-    }
+            val aepOptimizeError = AEPOptimizeError.getUnexpectedError()
+            OptimizeUtils.failWithOptimizeError(callback, aepOptimizeError)
+            return
+        }
 
-    /** Clears the client-side in-memory propositions cache. */
-    public static void clearCachedPropositions() {
-        OptimizeImpl.INSTANCE.clearCachedPropositions();
+        val validScopes = decisionScopes.filter { it.isValid() }
+
+        if (validScopes.isEmpty()) {
+            Log.warning(
+                OptimizeConstants.LOG_TAG,
+                SELF_TAG,
+                "Cannot update propositions, provided list of decision scopes has no valid scope."
+            )
+            return
+        }
+
+        val flattenedDecisionScopes = validScopes.map { it.toEventData() }
+
+        val eventData = mutableMapOf<String, Any>().apply {
+            put(
+                OptimizeConstants.EventDataKeys.REQUEST_TYPE,
+                OptimizeConstants.EventDataValues.REQUEST_TYPE_UPDATE
+            )
+            put(OptimizeConstants.EventDataKeys.DECISION_SCOPES, flattenedDecisionScopes)
+            xdm?.let { put(OptimizeConstants.EventDataKeys.XDM, it) }
+            data?.let { put(OptimizeConstants.EventDataKeys.DATA, it) }
+            put(
+                OptimizeConstants.EventDataKeys.TIMEOUT,
+                (timeoutSeconds * OptimizeConstants.TIMEOUT_CONVERSION_FACTOR).toLong()
+            )
+        }
+
+        val event = Event.Builder(
+            OptimizeConstants.EventNames.UPDATE_PROPOSITIONS_REQUEST,
+            OptimizeConstants.EventType.OPTIMIZE,
+            OptimizeConstants.EventSource.REQUEST_CONTENT
+        ).setEventData(eventData).build()
+
+        MobileCore.dispatchEventWithResponseCallback(
+            event,
+            (timeoutSeconds * OptimizeConstants.TIMEOUT_CONVERSION_FACTOR).toLong(),
+            object : AdobeCallbackWithError<Event> {
+                override fun fail(adobeError: AdobeError) {
+                    val aepOptimizeError = when (adobeError) {
+                        AdobeError.CALLBACK_TIMEOUT -> AEPOptimizeError.getTimeoutError()
+                        else -> AEPOptimizeError.getUnexpectedError()
+                    }
+                    OptimizeUtils.failWithOptimizeError(callback, aepOptimizeError)
+                }
+
+                override fun call(event: Event) {
+                    try {
+                        val eventData = event.eventData ?: run {
+                            OptimizeUtils.failWithOptimizeError(
+                                callback,
+                                AEPOptimizeError.getUnexpectedError()
+                            )
+                            return
+                        }
+
+                        if (eventData.containsKey(OptimizeConstants.EventDataKeys.RESPONSE_ERROR)) {
+                            val error = eventData[OptimizeConstants.EventDataKeys.RESPONSE_ERROR]
+                            if (error is Map<*, *>) {
+                                OptimizeUtils.failWithOptimizeError(
+                                    callback,
+                                    AEPOptimizeError.toAEPOptimizeError(error as Map<String, Any>)
+                                )
+                            }
+                        }
+
+                        if (!eventData.containsKey(OptimizeConstants.EventDataKeys.PROPOSITIONS)) return
+
+                        val propositionsList = DataReader.getTypedListOfMap(
+                            Any::class.java,
+                            eventData,
+                            OptimizeConstants.EventDataKeys.PROPOSITIONS
+                        )
+
+                        val propositionsMap = mutableMapOf<DecisionScope, OptimizeProposition>()
+                        propositionsList?.forEach { propositionData ->
+                            val optimizeProposition =
+                                OptimizeProposition.fromEventData(propositionData)
+                            if (optimizeProposition != null && !optimizeProposition.scope.isNullOrEmpty()) {
+                                val scope = DecisionScope(optimizeProposition.scope)
+                                propositionsMap[scope] = optimizeProposition
+                            }
+                        }
+
+                        callback?.call(propositionsMap)
+                    } catch (e: DataReaderException) {
+                        OptimizeUtils.failWithOptimizeError(
+                            callback,
+                            AEPOptimizeError.getUnexpectedError()
+                        )
+                    }
+                }
+            }
+        )
     }
 }
